pc104sg: More rework. A snapshot of the three clocks: Unix clock, IRIG registers, and our     10 msec ticker, is taken in the ISR once a second.  Then later in the 1 second callback     (software interrupt context) the ticker is compared against the hardware     clock, which is the IRIG clock if we have IRIG sync, or the Unix clock if no sync.     If the difference exceeds some limits, then the ticker is reset in the next ISR. This should     give us a little better estimate of the clock differences, and a better correction method.
    Added (I hope) better detection of missed interrupts, by checking the time deltas. 
    Should be able to #ifdef out the watchdog - will test that next.

    The 100 Hz counter, which indicates which of the 100,50,25,20,etc Hz callbacks are to
    be called, is 99% under the control of the bottom half, and is only reset by the
    ISR if things are really out of whack. A little period of no IRIG sync and clock wiggle
    should not then cause a 20 Hz callback (for example) from being called.
    The 100Hz counter is now an atomic type.

    1 second data records from IRIG contain two more 1 byte fields (rollover at 256 is OK):
    the number of missed interrupts, and the number of clock resets.

    1/10th millisecond clock ticker can be signed int - changed from unsigned in a few places.

nidas::core::SampleClock: Removed old cruft, which had been mostly #ifdef'd out since pre-PLOWS.
    The primary role of this class is to add date information to timetags which are an
    offset from 00:00 of the day. Old stuff in the this class tried to keep track of the
    IRIG-UNIX time difference, which was inaccurate since it was effected by the
    latency of an ioctl.

Other linux modules: changed ioport address types to long,
    in case we ever get a 64 bit ISA bus system (not likely, I hope)

+++ src/nidas/linux/diamond/dmd_mmat.c
@@ -2940,7 +2940,7 @@ static int __init dmd_mmat_init(void)
 
         for (ib = 0; ib < numboards; ib++) {
                 struct DMMAT* brd = board + ib;
-                unsigned long addr =  ioports[ib] + SYSTEM_ISA_IOPORT_BASE;
+                unsigned long addr =  (unsigned long)ioports[ib] + SYSTEM_ISA_IOPORT_BASE;
                 KLOG_DEBUG("isa base=%x\n",SYSTEM_ISA_IOPORT_BASE);
 
                 brd->num = ib;
